move logic to decider
