+++ Test/Unit/Helper/FeatureSwitch/ManagerTest.php
@@ -253,69 +253,4 @@ class ManagerTest extends TestCase
 
         $this->manager->updateSwitchesFromBolt();
     }
-
-    public function testIsSwitchEnabled_nothingInDbNoRollout() {
-        $this->fsRepo
-            ->expects($this->once())
-            ->method('getByName')
-            ->willThrowException(new NoSuchEntityException(__("no found")));
-
-        $fsVal = $this->manager->isSwitchEnabled(Definitions::M2_SAMPLE_SWITCH_NAME);
-
-        $this->assertEquals($fsVal, false);
-    }
-
-    public function testIsSwitchEnabled_ValInDbNoRollout() {
-        $fs = $this->createMock(FeatureSwitch::class);
-        $fs->expects($this->once())->method('getDefaultValue')->willReturn(true);
-        $fs->expects($this->once())->method('getRolloutPercentage')->willReturn(0);
-        $this->fsRepo
-            ->expects($this->once())
-            ->method('getByName')
-            ->willReturn($fs);
-
-        $fsVal = $this->manager->isSwitchEnabled(Definitions::M2_SAMPLE_SWITCH_NAME);
-
-        $this->assertEquals($fsVal, true);
-    }
-
-    public function testIsSwitchEnabled_ValInDbFullRollout() {
-        $fs = $this->createMock(FeatureSwitch::class);
-        $fs->expects($this->once())->method('getValue')->willReturn(true);
-        $fs->expects($this->exactly(2))
-            ->method('getRolloutPercentage')->willReturn(100);
-
-        $this->fsRepo
-            ->expects($this->once())
-            ->method('getByName')
-            ->willReturn($fs);
-
-        $fsVal = $this->manager->isSwitchEnabled(Definitions::M2_SAMPLE_SWITCH_NAME);
-
-        $this->assertEquals($fsVal, true);
-    }
-
-    public function testIsSwitchEnabled_ValInDbPartialRollout() {
-        $fs = $this->createMock(FeatureSwitch::class);
-        $fs->expects($this->once())->method('getValue')->willReturn(true);
-        $fs->expects($this->exactly(2))
-            ->method('getRolloutPercentage')->willReturn(100);
-
-        $this->fsRepo
-            ->expects($this->once())
-            ->method('getByName')
-            ->willReturn($fs);
-
-        $fsVal = $this->manager->isSwitchEnabled(Definitions::M2_SAMPLE_SWITCH_NAME);
-
-        $this->assertEquals($fsVal, true);
-    }
-
-    public function testIsSwitchEnabled_throwsIfBadSwitchName() {
-        $this->expectException(LocalizedException::class);
-        $this->expectExceptionMessage("Unknown feature switch");
-        $this->manager->isSwitchEnabled("SECRET_FEATURE_SWITCH");
-    }
-
-    // TODO(roopakv): Figure out how to mock globals to test rollout.
 }
